third commit

+++ src/main/java/com/blogging/employee/repositiories/EmployeeRepo.java
@@ -3,5 +3,8 @@ package com.blogging.employee.repositiories;
 import com.blogging.employee.entities.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepo extends JpaRepository<Employee, Integer> {
+import java.util.Optional;
+
+public interface EmployeeRepo extends JpaRepository<Employee, Long> {
+
 }
